change float to double; fixbug: cblas_dcopy, must use malloc to allocate memeory

+++ logistic_regression/include/data_struct.h
@@ -3,6 +3,6 @@
 
 struct sparse_feature{
     int idx;
-    float val;
+    double val;
 };
 #endif
